modified params type

+++ src/main/java/cn/edu/ldu/watercloud/entity/Watermeter.java
@@ -20,5 +20,5 @@ public class Watermeter {
     private double totalflow;
     private double lastflow;
     private double banalnce;
-    private double status;
+    private int status;
 }
